fix: prevent `ScanFile...` functions to consume too much memory when scanning large files (#42)
The previous implementation always read the whole file into a memory buffer before sending it to VirusTotal, this imposes a large memory footprint when scanning large files.

With the new implementation, we try to determine the payload size without reading the whole file. If the size can't be determined, then we fall back to reading the whole file into a buffer. In most cases reading the whole file is not required.

Closes #41.